se ha terminado l afuncion principal del carrito de compra
